% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Causality_Explained.R
\name{Causality_Explained}
\alias{Causality_Explained}
\title{Main function of Causality Explained}
\usage{
Causality_Explained(
  CutOff = 5e-08,
  DataPath,
  Outcomes,
  NBCores = 1,
  BonferroniCorrection = F,
  MinNbrIVs = 1
)
}
\arguments{
\item{CutOff}{which is the cutoff set to select IVs}

\item{DataPath}{folder containing the data (predictors, outcomes, information tables)}

\item{Outcomes}{outcomes to process}

\item{NBCores}{number of cores to run the pipeline in parallel}

\item{BonferroniCorrection}{if TRUE, apply a Bonferroni correction on predictor selection}

\item{MinNbrIVs}{The minimum number of significant IVs to use the trait as a predictor}
}
\description{
Main function of Causality Explained
}
